refactor: 가입 완료 페이지 멘티, 멘토 하나로 합침
