+++ src/components/templates/SignUpCompleteTemplate/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUpCompleteTemplate } from './SignUpCompleteTemplate';
