--- src/components/templates/SignUpMenteeCompleteTemplate/index.ts
@@ -1,1 +1,0 @@
-export { default as SignUpMenteeCompleteTemplate } from './SignUpMenteeCompleteTemplate';
