--- src/components/templates/SignUpMentorCompleteTemplate/index.ts
@@ -1,1 +1,0 @@
-export { default as SignUpMentorCompleteTemplate } from './SignUpMentorCompleteTemplate';
